Move database helpers to top level of package
