+++ collect_app/src/androidTest/java/org/odk/collect/android/instrumented/database/ChangingDatabasesByForksTest.java
@@ -3,8 +3,8 @@ package org.odk.collect.android.instrumented.database;
 import org.junit.Test;
 import org.odk.collect.android.dao.FormsDao;
 import org.odk.collect.android.dao.InstancesDao;
-import org.odk.collect.android.database.helpers.FormsDatabaseHelper;
-import org.odk.collect.android.database.helpers.InstancesDatabaseHelper;
+import org.odk.collect.android.database.FormsDatabaseHelper;
+import org.odk.collect.android.database.InstancesDatabaseHelper;
 import org.odk.collect.android.instrumented.database.helpers.SqlLiteHelperTest;
 
 import java.io.File;
